+++ doc/reference/CHANGELOG.md
@@ -13,6 +13,10 @@ All notable changes to this project will be documented in this file.
 ### Removed
 ### Fixed
 
+- Fix a few issues with the nodeAffinity setting logic, where the node affinity of a statefulset
+  would only be updated when its annotations are changed; and where it would hang if the statefulset's
+  replicas are set to 0.
+
 ## [v0.0.58] - 2025-09-05
 
 ### Added